Create 1009 Product of Polynomials
This time, you are supposed to find A×B where A and B are two polynomials.

Input Specification:

Each input file contains one test case. Each case occupies 2 lines, and each line contains the information of a polynomial:

K N
​1
​​  a
​N
​1
​​ 
​​  N
​2
​​  a
​N
​2
​​ 
​​  ... N
​K
​​  a
​N
​K
​​ 
​​ 

where K is the number of nonzero terms in the polynomial, N
​i
​​  and a
​N
​i
​​ 
​​  (i=1,2,⋯,K) are the exponents and coefficients, respectively. It is given that 1≤K≤10, 0≤N
​K
​​ <⋯<N
​2
​​ <N
​1
​​ ≤1000.

Output Specification:

For each test case you should output the product of A and B in one line, with the same format as the input. Notice that there must be NO extra space at the end of each line. Please be accurate up to 1 decimal place.